tehtävä

+++ suorakulmio.py
@@ -1,5 +1,5 @@
-kanta = int(input("Anna kanta: "))
-korkeus = int(input("Anna korkeus: "))
+kanta = float(input("Anna kanta: "))
+korkeus = float(input("Anna korkeus: "))
 piiri = kanta * 4
 korkeus = kanta * 2
 print(piiri)
